Bugfixes. Also some code refactoring.
After these bugfixes it seems to work pretty well, and
even outputs its item-weight.csv telemetry correctly
(most of that is probably changes in viking lib).

Refactoring in this commit is mostly the removal
of functions that are redundant because they have been
already moved into the library. Maybe a style consistency
tweak here and there.